+++ src/features/Product/components/Comment.module.css
@@ -1,0 +1,43 @@
+.comment {
+  border-bottom: 1px solid var(--gray200);
+  padding-bottom: 12px;
+}
+
+.contentContainer {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 24px;
+}
+
+.content {
+  font-size: 0.875rem;
+  font-weight: 400;
+  line-height: 1.5rem;
+  color: var(--gray800);
+}
+
+.authorInfo {
+  gap: 8px;
+}
+
+.authorInfo > div {
+  gap: 4px;
+}
+
+.authorInfo img {
+  width: 32px;
+  height: 32px;
+  border-radius: 50%;
+}
+
+.authorInfo span {
+  font-size: 0.75rem;
+  font-weight: 400;
+  line-height: 1.125rem;
+}
+
+.authorInfo time {
+  font-size: 0.75rem;
+  font-weight: 400;
+  line-height: 1.125rem;
+}
